Sensor_Ingestor/requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
uvloop==0.19.0
orjson==3.10.7
